requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1
httpx>=0.26.0
requests>=2.31.0
beautifulsoup4>=4.12.3
psutil>=5.9.8
apscheduler>=3.10.4
jinja2>=3.1.3
python-multipart>=0.0.9
playwright>=1.41.0
numpy>=1.26.4
pytest>=8.0.0
pytest-asyncio>=0.23.0
